account

+++ src/main/java/serviceBD/app/Controller/AccountController.java
@@ -65,7 +65,6 @@ public class AccountController {
     @ResponseBody
     public ResponseEntity<Account> gatAcc(@RequestBody Account account) throws GeneralSecurityException, IOException {
         Optional<Account> optionalAccount = accountRepository.findByUsername(account.getUsername());
-        //System.out.println(account.getUsername());
         if (optionalAccount.isPresent()) {
             Account acc = optionalAccount.get();
             byte[] salt = new String("12345678").getBytes();
@@ -73,7 +72,6 @@ public class AccountController {
             int keyLength = 128;
             String password= account.getPassword();
             SecretKeySpec key = createSecretKey(password.toCharArray(), salt, iterationCount, keyLength);
-           // String encryptedPassword = encrypt(password, key);
             String decryptedPassword = decrypt(acc.getPassword(), key);
             if (password.equals(decryptedPassword)) {
                 return new ResponseEntity<>(acc, HttpStatus.ACCEPTED);
@@ -97,7 +95,7 @@ public class AccountController {
 
         }
         else {
-        	int person_id =accountRepository.findPerson_id(id);
+        	int person_id = Math.toIntExact(accountRepository.findPerson_id(id));
         	accountRepository.deleteById(id);
         	personRepository.deleteById(person_id);
     }
